Made retrieving history more generic.
Previous version used to be C2SC specific so it would only work in that enviroment.
It now simply matches the starting digits per line, so it should work out of the box
for all Team Foundation Server instances. Makes it more generic, so users of the script
don't have to modify it before first use.